pyproject.toml
[tool.poetry]
name = "tradingnetwork"
version = "0.1.0"
description = ""
authors = ["Eugene Geym <geymzheka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^71.1.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
